requirements.txt
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.26.0
google-cloud-texttospeech==2.16.3
google-generativeai==0.5.2
google-search-results==2.4.2
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.49
langchain-google-genai==1.0.3
langchain-text-splitters==0.0.1
langsmith==0.1.53
PyAudio==0.2.14
picamera2==0.3.18
sounddevice==0.4.6
opencv-python
